% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pqCompress.R
\name{pqCompress}
\alias{pqCompress}
\title{Product Quantization Compression}
\usage{
pqCompress(data, num_subspaces, num_clusters)
}
\arguments{
\item{data}{A matrix of high-dimensional data points.}

\item{num_subspaces}{The number of subspaces.}

\item{num_clusters}{The number of clusters in each subspace.}
}
\value{
A list containing the centroids and the compressed data representation.
}
\description{
Compresses high-dimensional data using product quantization.
}
\examples{
data <- matrix(rnorm(1000), ncol = 100)
compressed <- pqCompress(data, 10, 5)
}
