% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pqSearchCpp.R
\name{pqSearchCpp}
\alias{pqSearchCpp}
\title{Product Quantization Search}
\usage{
pqSearchCpp(query, compressed_data, centroids)
}
\arguments{
\item{query}{查询向量。}

\item{compressed_data}{压缩后的数据。}

\item{centroids}{质心列表。}
}
\value{
返回搜索结果的索引。
}
\description{
这个函数执行基于乘积量化的搜索。
}
\examples{
data <- matrix(rnorm(1000), ncol = 100)
compressed <- pqCompress(data, 10, 5)
query <- rnorm(100)
nearest_neightbors <- pqSearchCpp(query, compressed$compressed_data, compressed$centroids)
}
